Правка аннотации

+++ maxapi/client/default.py
@@ -5,7 +5,7 @@ class DefaultConnectionProperties:
     Класс для хранения параметров соединения по умолчанию для aiohttp-клиента.
 
     Args:
-        timeout (int): Таймаут всего соединения в секундах (по умолчанию 5 * 30).
+        timeout (float): Таймаут всего соединения в секундах (по умолчанию 5 * 30).
         sock_connect (int): Таймаут установки TCP-соединения в секундах (по умолчанию 30).
         **kwargs: Дополнительные параметры, которые будут сохранены как есть.
 
@@ -14,12 +14,12 @@ class DefaultConnectionProperties:
         kwargs (dict): Дополнительные параметры.
     '''
 
-    def __init__(self, timeout: int = 5 * 30, sock_connect: int = 30, **kwargs):
+    def __init__(self, timeout: float = 5 * 30, sock_connect: int = 30, **kwargs):
         '''
         Инициализация параметров соединения.
 
         Args:
-            timeout (int): Таймаут всего соединения в секундах.
+            timeout (float): Таймаут всего соединения в секундах.
             sock_connect (int): Таймаут установки TCP-соединения в секундах.
             **kwargs: Дополнительные параметры.
         '''
